Fix segfault if graph have 2 many edge

+++ Galois-2.2.1/include/Galois/Graph/FileGraph.h
@@ -226,7 +226,7 @@ public:
   unsigned int size() const { return numNodes; }
 
   //! Returns the number of edges in the graph
-  unsigned int sizeEdges() const { return numEdges; }
+  uint64_t sizeEdges() const { return numEdges; }
 
   //! Returns the size of an edge
   size_t edgeSize() const { return sizeofEdge; }
